moved Logger to general utils dir
