+++ tools/sanity_checker/segmentation_checker.py
@@ -2,7 +2,7 @@ import numpy as np
 import supervisely_lib as sly
 from scipy import ndimage
 
-from similarity_scorer.utils.logger import Logger
+from utils.logger import Logger
 from .label_checker import LabelChecker, check_label_existence
