+++ tools/similarity_scorer/utils/feature_extractor.py
@@ -15,7 +15,7 @@ from PIL import Image
 from img2vec_pytorch import Img2Vec
 from tqdm import tqdm
 
-from .logger import Logger
+from utils.logger import Logger
 from .cache import Cache
 
 # Img2Vec
